Refactor layouts for public and app routes
