+++ src/app/(public)/layout.tsx
@@ -1,0 +1,3 @@
+export default function PublicLayout({ children }: { children: React.ReactNode }) {
+  return <>{children}</>;
+}
