+++ src/app/(public)/page.tsx
@@ -1,0 +1,1 @@
+export { default } from '@/components/home/LandingPage';
