--- src/app/app/page.tsx
@@ -1,1 +1,0 @@
-export { default } from '@/components/home/Home';
